+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+### Added
+- Add metadata to baselime mark command
 ### Fixed
 - Improved baselime iam output formatting
 ## [0.0.78] 2023-09-23